aws: Add example for custom HTTP client idle connection options
Adds example to the SDK for configuring custom HTTP client idle
connection keep alive options.
